use the master/outstation agnostic session to create a master/outstation agnostic TCP/TLS client task
