+++ dnp3/src/shared/errors.rs
@@ -1,5 +1,7 @@
 use crate::app::Shutdown;
 use crate::link::error::LinkError;
+use crate::util::phys::PhysLayer;
+use std::time::Duration;
 
 /// Communication sessions might be stopped due to being disabled, or due to being shutdown permanently
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
@@ -40,5 +42,49 @@ impl From<LinkError> for RunError {
 impl From<Shutdown> for RunError {
     fn from(_: Shutdown) -> Self {
         RunError::Stop(StopReason::Shutdown)
+    }
+}
+
+enum SessionType {
+    Master(crate::master::task::MasterTask),
+}
+
+/// Wrapper around a specific type of session (master or outstation)
+///
+/// Allows for the sharing of transport connectivity code (e.g. TCP/TLS client + serial)
+/// between masters and outstations
+pub(crate) struct Session {
+    inner: SessionType,
+}
+
+impl Session {
+    pub(crate) fn master(task: crate::master::task::MasterTask) -> Self {
+        Self {
+            inner: SessionType::Master(task),
+        }
+    }
+
+    pub(crate) async fn run(&mut self, io: &mut PhysLayer) -> RunError {
+        match &mut self.inner {
+            SessionType::Master(x) => x.run(io).await,
+        }
+    }
+
+    pub(crate) async fn shutdown(&mut self) {
+        match &mut self.inner {
+            SessionType::Master(x) => x.shutdown().await,
+        }
+    }
+
+    pub(crate) async fn wait_for_enabled(&mut self) -> Result<(), Shutdown> {
+        match &mut self.inner {
+            SessionType::Master(x) => x.wait_for_enabled().await,
+        }
+    }
+
+    pub(crate) async fn wait_for_retry(&mut self, duration: Duration) -> Result<(), StopReason> {
+        match &mut self.inner {
+            SessionType::Master(x) => x.wait_for_retry(duration).await,
+        }
     }
 }
