+++ dnp3/src/tcp/master.rs
@@ -1,15 +1,13 @@
-use std::time::Duration;
 use tracing::Instrument;
 
 use crate::app::{ConnectStrategy, Listener};
-use crate::app::{RetryStrategy, Shutdown};
 use crate::link::LinkErrorMode;
 use crate::master::task::MasterTask;
 use crate::master::{MasterChannel, MasterChannelConfig};
-use crate::shared::{RunError, StopReason};
+use crate::shared::Session;
+use crate::tcp::client::ClientTask;
 use crate::tcp::EndpointList;
-use crate::tcp::{ClientState, ConnectOptions, Connector, PostConnectionHandler};
-use crate::util::phys::PhysLayer;
+use crate::tcp::{ClientState, ConnectOptions, PostConnectionHandler};
 
 /// Spawn a task onto the `Tokio` runtime. The task runs until the returned handle, and any
 /// `AssociationHandle` created from it, are dropped.
@@ -44,7 +42,7 @@ pub fn spawn_master_tcp_client_2(
     listener: Box<dyn Listener<ClientState>>,
 ) -> MasterChannel {
     let main_addr = endpoints.main_addr().to_string();
-    let (mut task, handle) = MasterTcpTask::new(
+    let (mut task, handle) = wire_master_client(
         link_error_mode,
         endpoints,
         config,
@@ -62,107 +60,24 @@ pub fn spawn_master_tcp_client_2(
     handle
 }
 
-pub(crate) struct MasterTcpTask {
-    connector: Connector,
-    reconnect_delay: Duration,
-    task: MasterTask,
+pub(crate) fn wire_master_client(
+    link_error_mode: LinkErrorMode,
+    endpoints: EndpointList,
+    config: MasterChannelConfig,
+    connect_strategy: ConnectStrategy,
+    connect_options: ConnectOptions,
+    connection_handler: PostConnectionHandler,
     listener: Box<dyn Listener<ClientState>>,
-}
-
-impl MasterTcpTask {
-    pub(crate) fn new(
-        link_error_mode: LinkErrorMode,
-        endpoints: EndpointList,
-        config: MasterChannelConfig,
-        connect_strategy: ConnectStrategy,
-        connect_options: ConnectOptions,
-        connection_handler: PostConnectionHandler,
-        listener: Box<dyn Listener<ClientState>>,
-    ) -> (Self, MasterChannel) {
-        let (tx, rx) = crate::util::channel::request_channel();
-        let task = MasterTask::new(false, link_error_mode, config, rx);
-
-        let retry_strategy = RetryStrategy::new(
-            connect_strategy.min_connect_delay,
-            connect_strategy.max_connect_delay,
-        );
-
-        let task = Self {
-            connector: Connector::new(
-                endpoints,
-                connect_options,
-                retry_strategy,
-                connection_handler,
-            ),
-            reconnect_delay: connect_strategy.reconnect_delay,
-            task,
-            listener,
-        };
-        (task, MasterChannel::new(tx))
-    }
-
-    pub(crate) async fn run(&mut self) {
-        let _ = self.run_impl().await;
-        self.task.shutdown().await;
-        self.listener.update(ClientState::Shutdown).get().await;
-    }
-
-    async fn run_impl(&mut self) -> Result<(), Shutdown> {
-        loop {
-            self.listener.update(ClientState::Disabled).get().await;
-            self.task.wait_for_enabled().await?;
-            if let Err(StopReason::Shutdown) = self.run_connection().await {
-                return Err(Shutdown);
-            }
-        }
-    }
-
-    async fn run_connection(&mut self) -> Result<(), StopReason> {
-        loop {
-            self.run_one_connection().await?;
-        }
-    }
-
-    async fn run_one_connection(&mut self) -> Result<(), StopReason> {
-        self.listener.update(ClientState::Connecting).get().await;
-        match self.connector.connect().await {
-            Ok(phys) => {
-                self.listener.update(ClientState::Connected).get().await;
-                self.run_phys(phys).await
-            }
-            Err(delay) => {
-                tracing::info!("waiting {} ms to retry connection", delay.as_millis());
-                self.listener
-                    .update(ClientState::WaitAfterFailedConnect(delay))
-                    .get()
-                    .await;
-                self.task.wait_for_retry(delay).await
-            }
-        }
-    }
-
-    async fn run_phys(&mut self, mut phys: PhysLayer) -> Result<(), StopReason> {
-        match self.task.run(&mut phys).await {
-            RunError::Stop(s) => Err(s),
-            RunError::Link(err) => {
-                tracing::warn!("connection lost - {}", err);
-
-                self.listener
-                    .update(ClientState::WaitAfterDisconnect(self.reconnect_delay))
-                    .get()
-                    .await;
-
-                if !self.reconnect_delay.is_zero() {
-                    tracing::warn!(
-                        "waiting {} ms to reconnect",
-                        self.reconnect_delay.as_millis()
-                    );
-
-                    self.task.wait_for_retry(self.reconnect_delay).await?;
-                }
-
-                Ok(())
-            }
-        }
-    }
+) -> (ClientTask, MasterChannel) {
+    let (tx, rx) = crate::util::channel::request_channel();
+    let session = Session::master(MasterTask::new(false, link_error_mode, config, rx));
+    let client = ClientTask::new(
+        session,
+        endpoints,
+        connect_strategy,
+        connect_options,
+        connection_handler,
+        listener,
+    );
+    (client, MasterChannel::new(tx))
 }
